libvmon: tidy up some elastic buffer handling
Much of this stuff is pretty sloppy ever since it was first
written as a casual experiment.  Fixup types to use
size_t/ssize_t where appropriate, and free the actual realloc'd
member of the char_array struct - the container struct with the
assumption that the member is at the struct start.

+++ src/libvmon/vmon.h
@@ -49,14 +49,14 @@ typedef enum _vmon_proc_wants_t {
 /* we add some new complex types */
 typedef struct _vmon_char_array_t {
 	char	*array;
-	int	len;
-	int	alloc_len;
+	size_t	len;
+	size_t	alloc_len;
 } vmon_char_array_t;
 
 typedef struct _vmon_str_array_t {
 	char	**array;
-	int	len;
-	int	alloc_len;
+	size_t	len;
+	size_t	alloc_len;
 } vmon_str_array_t;
